Update README.md
Hi,
I am Olsam, a computer science graduate, software developer, mobile application and game developer.
We are a team of freelancers (organization named Olsamtech). We have 3+ years of experience in this field.
We are the trusted partner for some of the world’s leading enterprises and businesses. We have helped bring ideas to life while building sustainable business practices across industries. 
We build highly affordable custom software for companies large and small by applying modern design principles, in conjunction with the latest in cloud, mobile, and desktop technologies, we create tailored solutions that connect co-workers with each other and companies with their customers, simplify and accelerate business processes, and lower costs.